Inline actual test run logic into `Run` flag
